cardano-diffusion: moved NodeTo{Node,Client} modules
Moved to `ouroboros-network:cardano-diffusion` sublibrary.
