+++ decentralized-message-queue/decentralized-message-queue.cabal
@@ -88,7 +88,7 @@ library
     network-mux ^>=0.9,
     nothunks ^>=0.1.4 || ^>=0.2,
     optparse-applicative ^>=0.18,
-    ouroboros-network:{ouroboros-network, orphan-instances} ^>=0.22,
+    ouroboros-network:{ouroboros-network, cardano-diffusion, orphan-instances} ^>=0.22,
     ouroboros-network-api ^>=0.15,
     ouroboros-network-framework ^>=0.19,
     ouroboros-network-protocols ^>=0.15,
