+++ decentralized-message-queue/src/DMQ/Diffusion/Arguments.hs
@@ -15,11 +15,10 @@ import Control.Tracer (Tracer)
 import Network.DNS (Resolver)
 import Network.Socket (Socket)
 
-import DMQ.NodeToClient
-import DMQ.NodeToNode
+import DMQ.NodeToClient as NtC
+import DMQ.NodeToNode as NtN
 import DMQ.NodeToNode qualified as DMQ
 import Ouroboros.Network.Diffusion.Types qualified as Diffusion
-import Ouroboros.Network.NodeToNode (HandshakeTr)
 import Ouroboros.Network.PeerSelection.Churn (peerChurnGovernor)
 import Ouroboros.Network.PeerSelection.Governor.Types
            (ExtraGuardedDecisions (..), PeerSelectionGovernorArgs (..))
@@ -37,8 +36,8 @@ diffusionArguments
      , MonadTimer m
      , Exception exception
      )
-  => Tracer m (HandshakeTr ntnAddr NodeToNodeVersion)
-  -> Tracer m (HandshakeTr ntcAddr NodeToClientVersion)
+  => Tracer m (NtN.HandshakeTr ntnAddr)
+  -> Tracer m (NtC.HandshakeTr ntcAddr)
   -> Diffusion.Arguments
        () () () () () () ()
        exception
